+++ server/repository/src/migrations/v2_02_00/sync.rs
@@ -4,7 +4,7 @@ pub(crate) fn migrate(connection: &StorageConnection) -> anyhow::Result<()> {
     sql!(
         connection,
         r#"
-            ALTER TABLE sync_log ADD duration_in_seconds BIGINT DEFAULT 0 NOT NULL;
+            ALTER TABLE sync_log ADD duration_in_seconds INT DEFAULT 0 NOT NULL;
         "#,
     )?;
